extract "utility" and "errors" form __init__.py, find exact executables during check as side effect
