+++ extension/textext/requirements_check.py
@@ -1,4 +1,3 @@
-
 import abc
 import logging
 import os
@@ -9,43 +8,16 @@ import sys
 
 class Defaults(object):
     __metaclass__ = abc.ABCMeta
-    @abc.abstractproperty
-    def python(self): pass
 
     @abc.abstractproperty
-    def inkscape(self): pass
-
-    @abc.abstractproperty
-    def ghostscript(self): pass
-
-    @abc.abstractproperty
-    def pstoedit(self): pass
-
-    @abc.abstractproperty
-    def pdflatex(self): pass
-
-    @abc.abstractproperty
-    def lualatex(self): pass
-
-    @abc.abstractproperty
-    def xelatex(self): pass
-
-    @abc.abstractproperty
-    def console_colors(self):pass
+    def console_colors(self): pass
 
     @abc.abstractproperty
     def inkscape_extensions_path(self): pass
 
 
 class LinuxDefaults(Defaults):
-    python = "python2.7"
-    inkscape = "inkscape"
-    ghostscript = "ghostscript"
-    pstoedit = "pstoedit"
-    pdflatex = "pdflatex"
-    lualatex = "lualatex"
-    xelatex = "xelatex"
-    pdf2svg = "pdf2svg"
+
     console_colors = "always"
 
     @property
@@ -55,30 +27,6 @@ class LinuxDefaults(Defaults):
 
 class WindowsDefaults(Defaults):
 
-    def __init__(self):
-        # Append the location of our apps to the system path
-        sys.path.append("extension/textext")
-        import win_app_paths as wap
-
-        paths = os.environ.get('PATH', '').split(os.path.pathsep)
-        for result in [wap.get_pstoedit_dir(), wap.get_ghostscript_dir()]:
-            if result and result is not wap.IS_IN_PATH:
-                paths += [os.path.join(result)]
-
-        result = wap.get_imagemagick_command()
-        if result:
-            paths += [os.path.join(os.path.dirname(result))]
-
-        os.environ['PATH'] = os.path.pathsep.join(paths)
-
-    python = "python.exe"
-    inkscape = "inkscape.exe"
-    ghostscript = "gs.exe"
-    pstoedit = "pstoedit.exe"
-    pdflatex = "pdflatex.exe"
-    lualatex = "lualatex.exe"
-    xelatex = "xelatex.exe"
-    pdf2svg = "pdf2svg.exe"
     console_colors = "never"
 
     @property
@@ -87,7 +35,6 @@ class WindowsDefaults(Defaults):
 
 
 class LoggingColors(object):
-
     enable_colors = False
 
     def __call__(self):
@@ -129,13 +76,13 @@ class LoggingColors(object):
         BG_WHITE = "\033[107m"
 
         levels = [
-            VERBOSE,          # 5
-            logging.DEBUG,    # 10
-            logging.INFO,     # 20
+            VERBOSE,  # 5
+            logging.DEBUG,  # 10
+            logging.INFO,  # 20
             logging.WARNING,  # 30
-            logging.ERROR,    # 40
-            SUCCESS,          # 41
-            UNKNOWN,          # 42
+            logging.ERROR,  # 40
+            SUCCESS,  # 41
+            UNKNOWN,  # 42
             logging.CRITICAL  # 50
         ]
         names = [
@@ -159,16 +106,15 @@ class LoggingColors(object):
             BG_RED + FG_WHITE,
         ]
         if not LoggingColors.enable_colors:
-            colors = [""]*len(colors)
-            COLOR_RESET=""
+            colors = [""] * len(colors)
+            COLOR_RESET = ""
         return {name: (level, color) for level, name, color in zip(levels, names, colors)}, COLOR_RESET
 
 
 def set_logging_levels():
     level_colors, COLOR_RESET = get_levels_colors()
-    for name, (level,color) in level_colors.items():
+    for name, (level, color) in level_colors.items():
         logging.addLevelName(level, color + name + COLOR_RESET)
-
 
 
 class TrinaryLogicValue(object):
@@ -210,7 +156,7 @@ class TrinaryLogicValue(object):
 
 
 class RequirementCheckResult(object):
-    def __init__(self, value, messages, nested=None, is_and_node=False, is_or_node=False, is_not_node=False):
+    def __init__(self, value, messages, nested=None, is_and_node=False, is_or_node=False, is_not_node=False, **kwargs):
         self.value = TrinaryLogicValue(value)
         self.messages = messages
         self.nested = nested if nested is not None else []
@@ -219,6 +165,7 @@ class RequirementCheckResult(object):
         self.is_or_node = is_or_node
         self.is_not_node = is_not_node
         self.is_critical = None
+        self.kwargs = kwargs
 
     @property
     def color(self):
@@ -253,22 +200,22 @@ class RequirementCheckResult(object):
 
         if parent:
             if parent.is_and_node:
-                tail = parent.color+"/-and-"+self.color+nest_symbol+reset_color
+                tail = parent.color + "/-and-" + self.color + nest_symbol + reset_color
             elif parent.is_or_node:
-                tail = parent.color+"/--or-"+self.color+nest_symbol+reset_color
+                tail = parent.color + "/--or-" + self.color + nest_symbol + reset_color
             elif parent.is_not_node:
-                tail = parent.color+"/-not-"+self.color+nest_symbol+reset_color
+                tail = parent.color + "/-not-" + self.color + nest_symbol + reset_color
             else:
-                tail = parent.color+"/-----"+self.color+nest_symbol+reset_color
+                tail = parent.color + "/-----" + self.color + nest_symbol + reset_color
         else:
-            tail = self.color+nest_symbol+reset_color
+            tail = self.color + nest_symbol + reset_color
 
         if not parent:
             suffix = ""
         elif parent.nested[-1] is self:
             suffix = "      "
         else:
-            suffix = parent.color+"|"+reset_color+"     "
+            suffix = parent.color + "|" + reset_color + "     "
 
         if not self.messages:
             messages = [""]
@@ -283,7 +230,7 @@ class RequirementCheckResult(object):
 
             tail = suffix
         for nst in self.nested:
-            nst.print_to_logger(logger, offset + 1, prefix=prefix+suffix, parent=self)
+            nst.print_to_logger(logger, offset + 1, prefix=prefix + suffix, parent=self)
 
     def flatten(self):
         if len(self.nested) == 0:
@@ -293,36 +240,51 @@ class RequirementCheckResult(object):
             self.nested[i] = nst.flatten()
 
         if self.nested[0].is_or_node and self.is_or_node:
+            kwargs = dict(self.kwargs)
+            kwargs.update(self.nested[0].kwargs)
             return RequirementCheckResult(
                 self.value,
                 self.nested[0].messages + self.messages,
                 self.nested[0].nested + self.nested[1:],
-                is_or_node=True
+                is_or_node=True,
+                **kwargs
             )
 
         if self.nested[0].is_and_node and self.is_and_node:
+            kwargs = dict(self.kwargs)
+            kwargs.update(self.nested[0].kwargs)
             return RequirementCheckResult(
                 self.value,
                 self.nested[0].messages + self.messages,
                 self.nested[0].nested + self.nested[1:],
-                is_and_node=True
+                is_and_node=True,
+                **kwargs
             )
 
         if self.nested[-1].is_or_node and self.is_or_node:
+            kwargs = dict(self.kwargs)
+            kwargs.update(self.nested[-1].kwargs)
             return RequirementCheckResult(
                 self.value,
                 self.messages + self.nested[-1].messages,
                 self.nested[:-1] + self.nested[-1].nested,
-                is_or_node=True
+                is_or_node=True,
+                **kwargs
             )
 
         if self.nested[-1].is_and_node and self.is_and_node:
+            kwargs = dict(self.kwargs)
+            kwargs.update(self.nested[-1].kwargs)
             return RequirementCheckResult(
                 self.value,
                 self.messages + self.nested[-1].messages,
                 self.nested[:-1] + self.nested[-1].nested,
-                is_and_node=True
+                is_and_node=True,
+                **kwargs
             )
+
+        if self.nested[-1].is_not_node:
+            self.kwargs.update(self.nested[-1].kwargs)
 
         return self
 
@@ -343,6 +305,9 @@ class RequirementCheckResult(object):
             for nst in self.nested:
                 nst.mark_critical_errors(not non_critical_value)
 
+    def __getitem__(self, item):
+        return self.kwargs[item]
+
 
 class Requirement(object):
     def __init__(self, criteria, *args, **kwargs):
@@ -350,6 +315,10 @@ class Requirement(object):
         self._prepended_messages = {"ANY": [], "SUCCESS": [], "ERROR": [], "UNKNOWN": []}
         self._appended_messages = {"ANY": [], "SUCCESS": [], "ERROR": [], "UNKNOWN": []}
         self._overwrite_messages = None
+
+        self._on_unknown_callbacks = []
+        self._on_success_callbacks = []
+        self._on_failure_callbacks = []
 
     def check(self):
         result = self.criteria()
@@ -359,10 +328,16 @@ class Requirement(object):
         result.messages = self._prepended_messages["ANY"] + result.messages
         if result.value == TrinaryLogicValue(True):
             result.messages = self._prepended_messages["SUCCESS"] + result.messages
+            for callback in self._on_success_callbacks:
+                callback(result)
         if result.value == TrinaryLogicValue(False):
             result.messages = self._prepended_messages["ERROR"] + result.messages
+            for callback in self._on_failure_callbacks:
+                callback(result)
         if result.value == TrinaryLogicValue(None):
             result.messages = self._prepended_messages["UNKNOWN"] + result.messages
+            for callback in self._on_unknown_callbacks:
+                callback(result)
 
         result.messages += self._appended_messages["ANY"]
         if result.value == TrinaryLogicValue(True):
@@ -432,115 +407,187 @@ class Requirement(object):
 
         return Requirement(invert_impl)
 
+    def on_success(self, callback):
+        self._on_success_callbacks.append(callback)
+        return self
+
+    def on_failure(self, callback):
+        self._on_failure_callbacks.append(callback)
+        return self
+
+    def on_unknown(self, callback):
+        self._on_unknown_callbacks.append(callback)
+        return self
+
 
 class TexTextRequirementsChecker(object):
 
-    def __init__(self, logger):
+    def __init__(self, logger, config):
         self.logger = logger
+        self.config = config
+        self.available_tex_to_pdf_converters = {}
+        self.available_pdf_to_svg_converters = {}
+
+        self.inkscape_executable_name = "inkscape"
+        self.python27_executable_name = "python2.7"
+        self.pdflatex_executable_name = "pdflatex"
+        self.lualatex_executable_name = "lualatex"
+        self.xelatex_executable_name = "xelatex"
+        self.pdf2svg_executable_name = "pdf2svg"
+        self.pstoedit_executable_name = "pstoedit"
+        self.ghostscript_executable_name = "ghostscript"
+
+        self.inkscape_executable = None
+        self.python27_executable = None
+
+        self.pygtk_is_found = False
+        self.tkinter_is_found = False
+
         pass
 
-    def find_executable(self, executable_name):
-        messages = []
-        for path in os.environ["PATH"].split(os.path.pathsep):
-            full_path_guess = os.path.join(path, executable_name)
-            self.logger.log(VERBOSE, "Looking for `%s` in `%s`" % (executable_name, path))
-            if os.path.isfile(full_path_guess):
-                self.logger.log(VERBOSE, "`%s` is found at `%s`" % (executable_name, path))
-                messages.append("`%s` is found at `%s`" % (executable_name, path))
-        if len(messages) > 0:
-            return RequirementCheckResult(True, messages)
-        messages.append("`%s` is NOT found in PATH" % (executable_name))
-        return RequirementCheckResult(False, messages)
-
-    def find_PyGtk2(self):
+    def find_pygtk2(self):
         try:
-            subprocess.check_call([defaults.python, "-c", "import pygtk; pygtk.require('2.0'); import gtk;"],
+            executable = self.find_executable(self.python27_executable_name)["path"]
+            subprocess.check_call([executable, "-c", "import pygtk; pygtk.require('2.0'); import gtk;"],
                                   stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         except (OSError, subprocess.CalledProcessError):
             return RequirementCheckResult(False, ["PyGTK2 is not found"])
-
         return RequirementCheckResult(True, ["PyGTK2 is found"])
 
-    def find_TkInter(self):
+    def find_tkinter(self):
         try:
-            subprocess.check_call([defaults.python, "-c", "import TkInter; import tkMessageBox; import tkFileDialog;"],
+            executable = self.find_executable(self.python27_executable_name)["path"]
+            subprocess.check_call([executable, "-c", "import TkInter; import tkMessageBox; import tkFileDialog;"],
                                   stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         except (OSError, subprocess.CalledProcessError):
             return RequirementCheckResult(False, ["TkInter is not found"])
 
         return RequirementCheckResult(True, ["TkInter is found"])
 
-    def find_ghostscript(self, version):
+    def find_ghostscript(self, version=None):
         try:
-            p = subprocess.Popen([defaults.ghostscript, "--version"], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            executable = self.find_executable(self.ghostscript_executable_name)["path"]
+            p = subprocess.Popen([executable, "--version"], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
             stdout, stderr = p.communicate()
-        except (OSError, subprocess.CalledProcessError):
+        except (KeyError, OSError, subprocess.CalledProcessError):
             return RequirementCheckResult(False, ["ghostscript=%s is not found" % version])
+
+        if version is None:
+            return RequirementCheckResult(True, ["ghostscript is found"], path=executable)
 
         first_stdout_line = stdout.decode("utf-8").split("\n")[0]
         m = re.search(r"(\d+.\d+)", first_stdout_line)
         if m:
             found_version = m.group(1)
             if version == found_version:
-                return RequirementCheckResult(True, ["ghostscript=%s is found" % version])
+                return RequirementCheckResult(True, ["ghostscript=%s is found" % version], path=executable)
             else:
                 return RequirementCheckResult(False, [
                     "ghostscript=%s is not found (but ghostscript=%s is found)" % (version, found_version)])
         return RequirementCheckResult(None, ["Can't determinate ghostscript version"])
 
-    def find_pstoedit(self, version):
+    def find_pstoedit(self, version=None):
+
         try:
-            p = subprocess.Popen([defaults.pstoedit], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            executable = self.find_executable(self.pstoedit_executable_name)["path"]
+            p = subprocess.Popen([executable], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
             stdout, stderr = p.communicate()
-        except (OSError, subprocess.CalledProcessError):
+        except (KeyError, OSError, subprocess.CalledProcessError):
             return RequirementCheckResult(False, ["pstoedit=%s is not found" % version])
+
+        if version is None:
+            return RequirementCheckResult(True, ["pstoedit is found"], path=executable)
 
         first_stderr_line = stderr.decode("utf-8").split("\n")[0]
         m = re.search(r"version (\d+.\d+)", first_stderr_line)
         if m:
             found_version = m.group(1)
             if version == found_version:
-                return RequirementCheckResult(True, ["pstoedit=%s is found" % version])
+                return RequirementCheckResult(True, ["pstoedit=%s is found" % version], path=executable)
             else:
                 return RequirementCheckResult(False, [
                     "pstoedit=%s is not found (but pstoedit=%s is found)" % (version, found_version)])
         return RequirementCheckResult(None, ["Can't determinate pstoedit version"])
 
+    def find_executable(self, executable_name):
+        # try value from config
+        executable_path = self.config.get(executable_name+"-executable", None)
+        if self._check_executable(executable_path):
+            self.logger.debug("Using `%s-executable` from config = `%s`" % (executable_name, executable_path))
+            return RequirementCheckResult(True, "%s is found at `%s`" % (executable_name, executable_path), path=executable_path)
+        # look for executable in path
+        return self._find_executable_in_path(executable_name)
+
+    def _find_executable_in_path(self, executable_name):
+
+        messages = []
+        first_path = None
+        for path in os.environ["PATH"].split(os.path.pathsep):
+            full_path_guess = os.path.join(path, executable_name)
+            self.logger.log(VERBOSE, "Looking for `%s` in `%s`" % (executable_name, path))
+            if self._check_executable(full_path_guess):
+                self.logger.log(VERBOSE, "`%s` is found at `%s`" % (executable_name, path))
+                messages.append("`%s` is found at `%s`" % (executable_name, path))
+                if first_path is None:
+                    first_path = path
+        if len(messages) > 0:
+            return RequirementCheckResult(True, messages, path=os.path.join(first_path,executable_name))
+        messages.append("`%s` is NOT found in PATH" % (executable_name))
+        return RequirementCheckResult(False, messages)
+
+    def _check_executable(self, filename):
+        return filename is not None and os.path.isfile(filename) and os.access(filename, os.X_OK)
+
     def check(self):
 
+        def set_inkscape(exe):
+            self.inkscape_executable = exe
+
+        def add_latex(name, exe):
+            self.available_tex_to_pdf_converters.update({name: exe})
+
+        def set_pygtk(result):
+            self.pygtk_is_found = True
+
+        def set_tkinter(result):
+            self.tkinter_is_found= True
+
         textext_requirements = (
-                Requirement(self.find_executable, defaults.inkscape).prepend_message("ANY", 'Detect inkscape')
-                &
-                Requirement(self.find_executable, defaults.python).prepend_message("ANY", 'Detect `python2.7`')
-                &
+            Requirement(self.find_executable, self.inkscape_executable_name)
+            .prepend_message("ANY", 'Detect inkscape')
+            .on_success(lambda result: set_inkscape(result["path"]))
+            & Requirement(self.find_executable, self.inkscape_executable_name)
+            .prepend_message("ANY", 'Detect `python2.7`')
+            & (
+                    Requirement(self.find_executable, self.pdflatex_executable_name)
+                    .on_success(lambda result: add_latex("pdflatex", result["path"]))
+                    | Requirement(self.find_executable, self.lualatex_executable_name)
+                    .on_success(lambda result: add_latex("lualatex", result["path"]))
+                    | Requirement(self.find_executable, self.xelatex_executable_name)
+                    .on_success(lambda result: add_latex("xelatex", result["path"]))
+            ).overwrite_check_message("Detect *latex")
+            & (
+                    Requirement(self.find_pygtk2).on_success(set_pygtk) |
+                    Requirement(self.find_tkinter).on_success(set_tkinter)
+            ).overwrite_check_message("Detect GUI library")
+            & (
                 (
-                        Requirement(self.find_executable, defaults.pdflatex) |
-                        Requirement(self.find_executable, defaults.lualatex) |
-                        Requirement(self.find_executable, defaults.xelatex)
-                ).overwrite_check_message("Detect *latex")
-                &
-                (
-                        Requirement(self.find_PyGtk2) |
-                        Requirement(self.find_TkInter)
-                ).overwrite_check_message("Detect GUI library")
-                &
-                (
-                        (
-                                ~(
-                                        Requirement(self.find_pstoedit, "3.70") &
-                                        Requirement(self.find_ghostscript, "9.22")
-                                ).overwrite_check_message("Detect incompatible versions of psedit+ghostscript")
-                                &
-                                (
-                                        Requirement(self.find_executable, defaults.pstoedit) &
-                                        Requirement(self.find_executable, defaults.ghostscript)
-                                )
-                        ).overwrite_check_message("Detect compatible psedit+ghostscript versions")
-                        |
-                        (
-                            Requirement(self.find_executable, defaults.pdf2svg)
-                        ).prepend_message("ANY", "Detect pdf2svg:")
-                ).overwrite_check_message("Detect pdf->svg conversion utility")
+                    ~ (
+                            Requirement(self.find_pstoedit, "3.70") &
+                            Requirement(self.find_ghostscript, "9.22")
+                    ).overwrite_check_message("Detect incompatible versions of psedit+ghostscript")
+                    & (
+                            Requirement(self.find_pstoedit)
+                            .on_success(lambda result: self.available_pdf_to_svg_converters.update({"pstoedit": result["path"]}))
+                            & Requirement(self.find_ghostscript)
+                    )
+                ).overwrite_check_message("Detect compatible psedit+ghostscript versions")
+                .on_failure(lambda result: self.available_pdf_to_svg_converters.pop("pstoedit"))
+                | (
+                    Requirement(self.find_executable, self.pdf2svg_executable_name)
+                ).prepend_message("ANY", "Detect pdf2svg:")
+                .on_success(lambda result: self.available_pdf_to_svg_converters.update({"pdf2svg": result["path"]}))
+            ).overwrite_check_message("Detect pdf->svg conversion utility")
         ).overwrite_check_message("TexText requirements")
 
         check_result = textext_requirements.check()
@@ -560,7 +607,6 @@ if sys.platform.startswith("win"):
     defaults = WindowsDefaults()
 else:
     defaults = LinuxDefaults()
-
 
 VERBOSE = 5
 SUCCESS = 41
